+++ hmc/puppet/modules/hdp-oozie/manifests/download-ext-zip.pp
@@ -8,6 +8,7 @@ class hdp-oozie::download-ext-zip()
   
   exec{ "curl ${url}":
     command => "mkdir -p ${artifact_dir} ; curl -f --retry 10 ${url} -o ${target} ",
+    unless  => "test -f ${target}",
     creates => $target,
     path    => ["/bin","/usr/bin/"]
   }
