+++ hmc/puppet/modules/hdp-templeton/manifests/download-hive-tar.pp
@@ -8,6 +8,7 @@ class hdp-templeton::download-hive-tar()
   
   exec{ "curl ${url}":
     command => "mkdir -p ${artifact_dir} ; curl -f --retry 10 ${url} -o ${target} ",
+    unless  => "test -f ${target}",
     creates => $target,
     path    => ["/bin","/usr/bin/"]
   }
